+++ src/utils/sizeToTypography/index.ts
@@ -1,0 +1,9 @@
+import { Size } from "@ptypes/select/selectbusinessUnit/Istatus";
+
+const getTypo = (size: Size) => {
+  if (size === "compact") {
+    return "medium";
+  }
+  return "large";
+};
+export { getTypo };
